Add explicit type conversions for numpy2 compatibility

+++ grad_board.py
@@ -491,7 +491,7 @@ class GPAFHDO:
             grad_vals_cal = self.apply_cal(grad_vals, channel)
         else:
             grad_vals_cal = grad_vals
-        gr_dacbits = np.round(32767.51 * (grad_vals_cal + 1)).astype(np.uint16)
+        gr_dacbits = np.round(32767.51 * (grad_vals_cal + 1)).astype(np.uint32)
         gr = gr_dacbits | 0x80000 | (channel << 16)
 
         # # always broadcast for the final channel (TODO: probably not needed for GPA-FHDO, check then remove)
